+++ RMSDCalculator.cpp
@@ -31,26 +31,19 @@ int main(int argc, char *argv[])
     std::string IDs(argv[2]);
     std::transform(IDs.cbegin(), IDs.cend(), IDs.begin(), toupper);
 
-//     std::cout << "model->size" << model->size() << std::endl;
     //this contains how many mass points the chain has
     std::vector<int> chain_sizes;
     for(int i(0); i<model->size(); ++i)
     {
         chain_sizes.push_back(model->at(i)->get_size());
-//         std::cout<< "model->at(" << i << ")->get_size" << model->at(i)->get_size() << std::endl;
     }
-//     for(int i(0); i<chain_sizes.size(); ++i)
-//         std::cout << chain_sizes.at(i) << std::endl;
 
     //this contains what is the ID of the chain to calculate RMSD
     std::vector<int> chain_IDs;
     for(auto iter = IDs.begin(); iter != IDs.end(); ++iter)
     {
         chain_IDs.push_back(model->find_id(*iter));
-//         std::cout << "model->find_id" << model->find_id(*iter) << std::endl;
     }
-//     for(int i(0); i<chain_IDs.size(); ++i)
-//         std::cout << chain_IDs.at(i) << std::endl;
 
     std::string dcd(filename + ".dcd");
     DCDReader dcdreader;
@@ -78,10 +71,6 @@ int main(int argc, char *argv[])
     SnapShot init(pickup_chain(initial.first, chain_IDs, chain_sizes));
     SnapShot seco(pickup_chain(initial.first, chain_IDs, chain_sizes));
 
-//     std::cout << "init size " << init.size() << std::endl;
-//     std::cout << "seco size " << seco.size() << std::endl;
-
-
     rmsdcalc.set_data(init, init);
     ofs << 0e0 << " " << rmsdcalc.get_RMSD() << std::endl;
 
@@ -92,7 +81,6 @@ int main(int argc, char *argv[])
     {
         std::pair<SnapShot, double> snapshot
             (dcdreader.get_snapshot(i));
-//         SnapShot sshot(snapshot.first.begin(), (snapshot.first.begin()+end_ID));
         SnapShot sshot(pickup_chain(snapshot.first, chain_IDs, chain_sizes));
 
         rmsdcalc.set_data2(sshot);
@@ -110,7 +98,7 @@ SnapShot pickup_chain(const SnapShot& ss,
 
     SnapShot::iterator ssiter = retval.begin();
 
-    for(int i(0); i<chain_sizes.size(); ++i)
+    for(size_t i(0); i<chain_sizes.size(); ++i)
     {
         if(std::find(chain_ids.begin(), chain_ids.end(), i) != chain_ids.end())
         {
